requirements.txt
langchain
sentence-transformers
python-dotenv>=0.19.0
PyQt5>=5.15.0
openai>=0.27.0
requests==2.31.0
Pillow
python-dotenv==1.0.0
playwright>=1.41.2
pyee>=11.0.1
pyppeteer-stealth>=2.7.4
azure-cognitiveservices-speech
